backend/requirements.txt
astroid==2.12.2
Click==8.1.3
ecdsa==0.18.0
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
future==0.18.2
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.0
MarkupSafe==2.1.1
mccabe==0.7.0
pycryptodome==3.15.0
pylint==2.14.4
python-jose-cryptodome==1.3.2
six==1.16.0
typed-ast==1.4.2
Werkzeug==2.1.2
wrapt==1.14.1
Flask-Cors==3.0.10
